requirements.txt
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
bcrypt==4.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
cssselect2==0.7.0
Django==4.1.3
django-ninja==0.19.1
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
iniconfig==1.1.1
lxml==4.9.2
msgpack==1.0.4
oscrypto==1.3.0
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF3==1.0.6
pytest==7.2.0
python-bidi==0.4.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
Random-Word==1.0.11
reportlab==3.6.12
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.13
webencodings==0.5.1
xhtml2pdf==0.2.8
python-dotenv~=1.0.0